Добавлен класс-описание завода по производству дисков, обновлена форма создания заказа работы для завода

+++ source/monitor/NrpComponentListbox.cpp
@@ -562,7 +562,7 @@ void CNrpComponentListbox::draw()
 						
 						swprintf( tmpstr, 127, L"%s  (%d %%)", name.c_str(), (int)(percent * 100) );
 						core::recti progressRect = frameRect;
-						progressRect.LowerRightCorner.X = progressRect.UpperLeftCorner.X + frameRect.getWidth() * percent;
+						progressRect.LowerRightCorner.X = (s32)(progressRect.UpperLeftCorner.X + frameRect.getWidth() * percent);
 						driver->draw2DRectangle( progressRect, 0xff0000ff, 0xff0000ff, 0xff00ff00, 0xff00ff00, &clientClip );
 						Font->draw( tmpstr, textRect, itbncolor, false, true, &clientClip );
 					}
